+++ app/models/topic.rb
@@ -13,6 +13,8 @@ class Topic < ApplicationRecord
   }
   DEFAULT_ORDER = "topics.pinned desc, topics.last_reply_at desc, topics.id desc"
 
+  AUTO_LOCK_IN = 2.months
+
   attr_accessor :description
 
   # Concerns
@@ -31,6 +33,9 @@ class Topic < ApplicationRecord
 
   # Scopes
   scope :shadow_banned, ->(arg) { joins(:user).merge(User.where(shadow_banned: [nil, false]).or(User.where(id: arg))) }
+  # Auto-lock after one month.
+  scope :not_auto_locked, -> { where(locked: false).where("DATE(last_reply_at) >= ?", Time.zone.today - AUTO_LOCK_IN) }
+  scope :auto_locked, -> { where(locked: true).or(where("DATE(last_reply_at) < ?", Time.zone.today - AUTO_LOCK_IN)) }
 
   # Validations
   validates :title, presence: true
@@ -87,7 +92,7 @@ class Topic < ApplicationRecord
   end
 
   def editable_by?(current_user)
-    !locked? && (user == current_user || current_user.admin?)
+    (!auto_locked? && user == current_user) || current_user.admin?
   end
 
   def compute_shadow_ban!
@@ -112,7 +117,15 @@ class Topic < ApplicationRecord
   end
 
   def subscribed?(current_user)
-    current_user.subscriptions.where(topic_id: id, subscribed: true).count > 0
+    current_user.subscriptions.where(topic_id: id, subscribed: true).count.positive?
+  end
+
+  def auto_locked?
+    locked? || last_reply_at.to_date < Time.zone.today - AUTO_LOCK_IN
+  end
+
+  def not_auto_locked?
+    !auto_locked?
   end
 
   private